requirements.txt
boto3==1.40.52
botocore==1.40.52
datasets==4.1.1
elasticsearch==9.1.1
fastapi==0.119.0
httpx==0.28.1
langchain==0.3.27
langchain_community==0.3.31
langchain_google_genai==2.1.12
langchain_text_splitters==0.3.11
pandas==2.3.3
passlib==1.7.4
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
psycopg2_binary==2.9.10
pydantic==2.12.2
PyJWT==2.10.1
pyotp==2.9.0
pytest==8.4.2
python-dotenv==1.1.1
python_bcrypt==0.3.2
ragas==0.3.6
unstructured==0.18.15
upstash_redis==1.4.0
uvicorn==0.37.0
